* updated to v0_4_7 * updated timeouts - default to 180 sec and max to 1 hour * fixed stop in case of mismatch with source lang during import * Updated logs * Code cleaanup * enabled t5memory_loglevel * fixed issue with big AllowedMemory values * fixed wrong url parsing when it's not ends with slash * fixed -136

+++ source/RestAPI/OtmMemoryServiceWorker.cpp
@@ -302,7 +302,7 @@ int OtmMemoryServiceWorker::getFreeSlot(size_t memoryRequested)
   size_t UsedMemory = 0;
   int AllowedMemoryMB;
   properties_get_int(KEY_ALLOWED_RAM, AllowedMemoryMB);
-  size_t AllowedMemory = AllowedMemoryMB * 1000000;
+  size_t AllowedMemory = (size_t)AllowedMemoryMB * 1000000;
 
   #ifdef CALCULATE_ONLY_MEM_FILES
   std::string path;
